code/docker/3-concours/requirements.txt
django
celery
djangorestframework